added tech attribute to the projects and updated animations
