+++ content/projects/project4.json
@@ -4,6 +4,10 @@
   "shortDescription": "A browser-based puzzle game that traces the history of web dev tech by having you connect evolution chains across multiple timelines.",
   "cover": "/images/Evolution_menu.png",
   "type": "Course Project",
-  "technologies": ["HTML5", "CSS3", "JavaScript"],
+  "tech": [
+    "HTML5",
+    "CSS3",
+    "JavaScript"
+  ],
   "githubLink": "https://github.com/Ph4ntomByte/Evolution-Game"
 }
